requirements.txt
# MCP and core dependencies
mcp>=0.5.0
pydantic>=2.0.0
python-dotenv>=1.0.0
aiohttp>=3.8.0

# News and AI APIs
eventregistry>=1.2.6
google-generativeai>=0.3.0

# Publishing platforms
PyGithub>=1.59.0
tweepy>=4.14.0

# Utilities
python-dateutil>=2.8.0
markdown>=3.4.0
jinja2>=3.1.0

# Development
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
